requirements.txt
Django>=2.1.3,<2.2.0
djangorestframework>=3.9.0,<3.10.0
psycopg2>=2.7.5,<2.8.0
Pillow>=5.3.0,<5.4.0

flake8>=3.6.0,<3.7.0
coverage>=4.4.2,<4.5.4
coveralls>=1.7.0,<1.8.2
black==19.10b0